new adwords report source (budget perf), fix average cpv and update func

+++ src/config/adwords/sources/ad/averagecpv.php
@@ -7,8 +7,8 @@ return [
     "fields" => [
         "AverageCpv"
     ],
-    "parse" => function ($data): int {
-        return (int)$data->AverageCpv;
+    "parse" => function ($data): float {
+        return (float)$data->AverageCpv;
     },
     "inferred_from" => [
         "cost",
